Took out old rendering formats.
Decreases size of executable, which is good, at the expense of a few old
replays we won't be able to read any more through the main version.
I've copied the old executable, however, to allow  the occasional
browsing of old games as one might wish, but we'll overall move away
from that.